Started Work on Weekly Lessons page (untested)
Added two new database methods to get filtered database results (untested)
Added action button that pushes a new dialog with the current week and next week.
Current day is highlighted, and calendar always starts at the last sunday.
Added potential support for a list of lessons for each day, as well as the name of the student taking it.