Implemented Manoeuvre selection list when editing an existing lesson.
In the new dialog a list of checkboxes listing all compatible manoeuvres (ones with the same category as the current lesson, as well as ones present in the lesson but not in the database) can be selected and enabled/disabled. Disabling a non-existent manoeuvre will remove it from the lesson permanently.
//TODO: Possibly make it so that when a manoeuvre is renamed, the app goes through all existing lessons and changes the name of the manoeuvre, if it has it, to the new name. Do the same for categories.
What happens when a category is deleted? Delete students? Or keep them, like I do for delete manoeuvres?

Dynamic Marquee listing a lesson's manoeuvres now actually lists them, instead of a placeholder text. The string is filtered to make sure blank spaces (unnecessary line breaks) are left out.