Implemented Lesson Details Dialog
When the user presses a specific lesson (its button on the lesson list), this dialog shows up.
It is an adaptation of the edit lesson dialog, geared towards static data. The data is displayed in a similar form to the Manoeuvre and Category Details Dialogs.

Changed visual lesson completion indication on the lesson list, so it does not resemble a checkbox. This tells the user it is not a clickable UI element.