Chand the previous janky and unextensible system into a new, safer one based on a hybrid class
This class contains both arguments from lessons and exams, as well as a boolean that identifies which type it is. This way, both are converted into the same type of object, whilst maintaining their exclusive attributes. This also has the advantage of making compatible code easier to write, as we do not need to keep track of two different types.